+++ pre_commit_hooks/check_jamf_json_schemas.py
@@ -1,0 +1,39 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+"""This hook checks Jamf JSON schemas for inconsistencies and common issues."""
+
+import argparse
+import json
+
+
+def build_argument_parser():
+    """Build and return the argument parser."""
+
+    parser = argparse.ArgumentParser(
+        description=__doc__, formatter_class=argparse.RawDescriptionHelpFormatter
+    )
+    parser.add_argument("filenames", nargs="*", help="Filenames to check.")
+    return parser
+
+
+def main(argv=None):
+    """Main process."""
+
+    # Parse command line arguments.
+    argparser = build_argument_parser()
+    args = argparser.parse_args(argv)
+
+    retval = 0
+    for filename in args.filenames:
+        try:
+            with open(filename, "rb") as openfile:
+                schema = json.load(openfile)
+        except Exception as err:
+            print("{}: json parsing error: {}".format(filename, err))
+            retval = 1
+
+    return retval
+
+
+if __name__ == "__main__":
+    exit(main())